refactor(App): Move the functions of the App class to the LottoView class
App 클래스의 기능을 LottoView 클래스로 이동
