+++ docs/README.md
@@ -45,8 +45,8 @@
 - [x] 오름 차순으로 출력
 - [x] 보너스 번호 입출력 기능
 - [x] 당첨내역 출력
-- [ ] 수익률 출력(소수점 둘째 자리에서 반올림)
-- [ ] 예외 사항 시 에러 출력("[ERROR]로 시작")
+- [x] 수익률 출력(소수점 둘째 자리에서 반올림)
+- [x] 예외 사항 시 에러 출력("[ERROR]로 시작")
 
 ### 입력 유효성 검사
